Fix Node installation script
Previously the Node install process was broken: it attempted a global
install to `/usr/local/nvm` but was unable to create the directory.

In practice, a local install in the user's home directory is probably
fine, and makes the process much simpler.

To get the install process to work smoothly (making `nvm` and `node`
seamlessly available without having to open a new terminal, it seems
to be necessary to run the script via `source` rather than `bash`.
As I understand it, this allows environment variables to persist
beyond the script in which they were set (so if the node setup script
makes `node` available, it remains available to subsequent scripts and
also on the command line subsequently).

While doing this, I also made the setup directory a variable, to avoid
a bit of repetition and make it easier to change in future.