Fix Ruby installation script
As with PHP, the `apt` install steps here need to be sudo'ed.
Additionally, `rbenv` requires a C compiler to run, so I needed
to add the `build-essential` package to fix this.

Finally, I removed the `-p` flag on `mkdir` to prevent one (non-
fatal) error when it tried to create the `$(rbenv root)` directory
(which at that point already exists). In fact, this command only
needs to create the `/plugins` sub-directory, so there's no need
for the `-p` flag.