Add .gemrc file with `no-document` option
Installing documentation alongside gems can take up a lot of space and slow
down installation. I don't ever use it, so no point in installing it.

As well as adding the `.gemrc` file, this commit also adds a line to the
`install.sh` script that will create a symlink to the .dotfiles version in
the home directory if `.gemrc` doesn't already exist.