Remove conditional from yarn setup script
Previously this conditional meant the script would not do anything if a
version of yarn was already present. However, this isn't necessarily
desirable: this script will install the latest version of yarn, and even
a long out-of-date version would prevent it running. This commit removes
the conditional so that it can be run even if yarn is present.