Add build-essential install step to new machine setup script
The `build-essential` package installs `make` and other tools needed to
build other packages. Previously it was being included as a dependency
in the Ruby setup script. However, it's also needed by some Node
packages, so it seems better to just install it by default.

This commit adds an initial step to install `build-essential` if `make`
is not found on the host machine.