Add an install prompt before installing command line tools
Previously the last few things in the "setup-new-machine" script just
ran automatically. However, it's better (and clearer) to prompt for
whether or not to install these, as we do for all the other stuff.

This commit wraps the install step for, eg. `fzf`, `tldr`, etc, in a
yes/no install prompt.