Add alias for "open" / "start"
`open` on OSX / `start` in Windows open a file in the default
application. This is useful to be able to do, and can be achieved in WSL
by calling `start` via `cmd.exe`[1]. This commit aliases this command to
both `open` and `start` (since tutorials, etc, are more likely to refer
to `open`).

[1] https://superuser.com/a/1252579/15645