Make sure rbenv's ruby-build plugin is up-to-date before updating Ruby
In order for rbenv to be able to update Ruby to the latest version, I
first had to update the `ruby-build` plugin. This commit adds that step
to the setup script for Ruby so that in future it should be possible to
just bump the Ruby version and `source` the script again in order to
update Ruby.