Update `rsc` alias to not run when no modified spec files are present
Previously if run with no modified spec files this would end up just
calling `rspec` with no args and therefore running the entire test
suite. Adding the `--no-run-if-empty` flag to `xargs` prevents this.

This command is still slightly constrained by only running on modified
specs. It's a little fiddly to update this to deal with newly-created
specs too, since instead of "modified:", Git indicates these with "new
file:" which (because of the space) makes `awk` split the string
differently (so $2 refers to the word "file" instead of the
spec filename).