Add Docker setup script
This installs the Docker Engine in WSL, which provides CLI tools for
interacting with Docker. To use Docker locally, you still need to
install Docker for Windows: the stuff installed here just allows you to
interact with the Docker daemon from the WSL command line.

Probably when WSL2 comes out (later this year), it will be possible to
install Docker itself directly in WSL, so the process of using it will
become a bit easier. For now, this seems to be the simplest way to use
Docker in WSL.

The steps in this script are mostly based on this post [1] and the
instructions in the Docker docs [2].

This way of using Docker does involve exposing a port without TLS, which
is a possible attack vector. So use with some caution. Hopefully this is
just a stop-gap till WSL2 comes out...

[1] https://davidburela.wordpress.com/2018/06/27/running-docker-on-wsl-windows-subsystem-for-linux/
[2] https://docs.docker.com/install/linux/docker-ce/ubuntu/