Prepend n to PATH
When n is installed, it appends its directory to the PATH. This can
present a problem with WSL, because WSL also (by default, anyway)
imports the PATH variable from Windows. Which means that if Node is
installed in Windows, it takes precedence over the version installed
via `n`

Prepending n to the PATH in WSL ensures that the version of Node
installed via `n` takes precedence.

I haven't attempted to remove the line that `n-install` adds, so as a
result of this this line does now get both prepended and appended.
I don't think that should really cause any problems though.