Add setup script for marker
To quote its README [1]:

"Marker is a command palette for the terminal. It lets you bookmark
commands (or commands templates) and easily retreive them with the help
of a real-time fuzzy matcher."

By default the key binding for marker is Ctrl-Space. However, this seems
to cause issues in Ubuntu so I've followed this suggestion [2] and
side-stepped this issue by binding marker to Ctrl-X instead.

[1] https://github.com/pindexis/marker
[2] https://github.com/pindexis/marker/issues/45#issuecomment-483928442