Use MySQL password rather than socket plugin for root user
By default MySQL uses the "unix_socket" plugin for root user access.
This means that `sudo` is required for root access. For development
purposes I'm not sure this is necessary, so let's switch to
"mysql_native_password" instead, which allows non-sudo login as the
root user.

It's worth noting that this is not at all secure (especially because
the root password might be unset), but for a local development DB it's
probably fine, and avoids confusion when trying to log in after
installing MySQL with the `-y` flag (ie. avoiding manual configuration)