Add shorthand aliases to gitconfig
For a long time I've used bash aliases for various git commands.
This works fine, but the advantage of setting aliases in git is that
things like autocompletion work. So let's add some aliases here instead.

I'm not sure if there are any conventions here: if there are it might
make sense to use them, but for now I've just made all these up.