Use "git ch" as shorthand for "git checkout" instead of "git co"
Although I've used "gco" as a shortcut for "git checkout" for ages,
on reflection I think "git co" is probably best avoided, because it's
ambiguous: it could easily stand for "git commit", and in a situation
where you intend to commit, accidentally doing "checkout" instead could
lead to loss of work.

So let's avoid "co" and go for the less ambiguous "ch" instead.