Add "git wip" alias
This is a series of actions I often take when changing branches
(ie. commit all work-in-progress to the current branch, bypassing any
linters via the "--no-verify" flag), so let's make create a shorthand
for it.

nb. the "!" prefix is needed to allow a git alias to run multiple
commands or external (ie. non-Git) commands [1][2]

[1] https://git-scm.com/book/en/v2/Git-Basics-Git-Aliases
[2] https://stackoverflow.com/a/18860775/180500