Add "git br" alias for nicely formatted list of branches
For a long time I've been using an alias of "gbr" that does more-or-less
this, but with the output piped to `column` to format it neatly.

However, in the latest version of Ubuntu `column` seems to cause the
colours to be lost, presumably for some reason to do with ANSI
colour codes [1].

I haven't figured out how to fix this, so in the mean time, rather than
use `column`, let's just use a pair of spaces and preserve the colours.

[1] https://unix.stackexchange.com/questions/251751/issue-with-column-command-and-color-escape-codes