Create Ansible playbook and bootstrap script
This commit introduces a bootstrap script that will install Ansible, if
necessary, and then run the playbook in `ansible/playbook.yml`

For now the playbook just uses `apt` to install some useful tools.

In subsequent commits, I'll move more of the setup scripts into Ansible,
and remove them from `setup`. I might leave some of the more optional
things in `setup` for now if I don't want them to always be installed.

Why use Ansible?

The main reason is that it's idempotent: ie. it can be run multiple
times and will always leave things in the same state. So if something's
already installed running the installer again won't mess with the
existing installation.

I've tried to write the bash setup scripts in a moderately idempotent
way (checking for the presence of things before performing certain
actions, etc), but it's hard to be sure I've always got this right.

Using Ansible makes it much easier to configure things in a declarative
way: I want this and this and this installed, but I don't necessarily
need to specify how (or whether) to install them. Ansible will handle
that for me.