Ansible: install "n" version manager, plus latest version of NodeJS
I don't want the latest version of NodeJS to be installed every time
this playbook runs. Instead, let's just install "n" with the optional
"lts" flag to install whatever the LTS version of Node is at the time
that "n" installs. Once "n" is installed, I'll use that to manage my
local versions of Node.

n's installer is a bit picky about existing Node versions: if Node is
installed in Windows it will refuse to install n in WSL, because
the Windows version of npm shows up in $PATH. In practice, this
shouldn't matter to n, but in order to force it to install properly
we have to do this hacky thing were we override $PATH so it doesn't
see the non-WSL version of npm. /bin and /usr/bin were both needed
in order for `bash` to run the install script successfully.