Ansible: install some npm packages
Here are a few things I find handy.
Now Node is installed via Ansible, let's install these this way too.

Note that in order to use n's version of npm, we need to tell Ansible
where to find the npm executable (`~/n/bin/npm`)

We're already setting `core.pager` to `diff-so-fancy` in `.gitconfig`,
so let's just delete the line that sets that from `setup-new-machine.sh`