Ansible: install yarn
It seems that nowadays yarn's own installation guide suggests using
npm to install yarn. This suits me fine, because the yarn PPA repo
was a bit unreliable and an occasional source of problems.

This might also have the advantage that, by tying yarn to a particular
version of node, I'll have to reinstall it when I change node versions.