Move everything from base.sh into .bashrc
In many ways base.sh served the same purpose as .bashrc, so let's
combine them. On reflection I think it's slightly confusing to have
base.sh as a shell script: it looks like something that might be run
from the command line, whereas it's really just an include. So let's
move its contents into .dotfiles/.bashrc and get rid of it.

Now the only ".sh" files are scripts that are actually intended to
be run from the command line.