Rename "setup-new-machine.sh" to "setup-extras.sh"
At this point all the essentials are handled by bootstrap.sh,
which can be reliably run at any point to bring a machine up to date.

"setup-extras.sh" is now just a convenience tool for running one or
more of the remaining setup scripts in ./setup for non-essential things
that may be useful to have.

At some point, I could move all this stuff into a different Ansible
role (eg. "extras"), but I'm not sure it's worth the effort for now.