Merge install.sh into bootstrap.sh
Originally, "install.sh" was a script you run once when you first
install the dotfiles, and "bootstrap.sh" was a script that could
be run any time to bring things up to date.

Now Ansible is responsible for most of the install process, there's no
real need for a script that only gets run once: let's combine it all
into a single bootstrap script that can be run any time.