Git: change default pull behaviour from "preserve" to "merges"
`git pull --rebase=preserve` does a `git pull -r` on the current branch:
ie. pull in the remote version of the branch and rebase the local
version on top of it using the `--preserve-merges` strategy, preserving
any merge commits (since rebase by default drops merge commits).

However, rebase's `--preserve-merges` strategy is being deprecated in
favour of `--rebase-merges`, which is very similar but works better with
interactive rebases [1].

Let's update this config setting to fix the deprecation warnings it
causes in the latest version of Git.

[1] https://git-scm.com/docs/git-rebase