Install asdf using git rather than a shell command
I hadn't realised, but it turns out that Ansible has a "git" module
that allows you to clone a repo. Let's use that rather than running
`git clone` as a shell command.

(fzf is still installed in the old way, but in that case the command
also runs the install script, etc - doesn't seem worth the effort
of rejigging)