Allow Ansible to only run specific tasks
Setting a `--tags` flag when running `ansible-playbook` allows you to
run only tasks that have that tag. This is quite handy for testing
playbooks, but is also nice to have as an option. I haven't actually
added any tags yet, but this change provides a reason to add them.

This is an idea I spotted in these dotfiles:
https://github.com/kespinola/dotfiles/blob/master/bin/dot-bootstrap

...and which I've borrowed because it looks handy and my bootstrap
script is otherwise pretty similar.