Do not add PPA repository for Ansible
This is no longer needed in Ubuntu 20.04, and in that case the presence
of this repo causes problems when running `apt-get update`

It doesn't seem worth adding a version check, but I'll leave this line
commented out for now, for ease of uncommenting if I do need to run
this on an older version of Ubuntu. At some point this line should
just be removed.