Set ansible_python_interpreter to "auto"
By default Ansible currently uses /usr/bin/python if present, even if
that's Python 2, for backwards compatibility reasons. When it does this
it outputs a deprecation warning. Setting `ansible_python_interpreter`
to "auto" tells Ansible to use whichever version it thinks is most
appropriate. This is the future default, and fixes the deprecation
warning too.