Add "serve" for quickly serving local folders
Sometimes it's useful to be able to quickly start up a web server to
serve files from the current directory.

That's what `serve` does. I've also used `http-server` in the past,
which is the same sort of thing but seems to have a major dependency
that's no longer maintained [1], so `serve` seems like a better bet.

`ngrok` can also do this [2], but it's fiddlier than just typing `serve`

[1] https://github.com/http-party/http-server/issues/619
[2] https://ngrok.com/docs#http-file-urls