Use local git completions file, not dotfiles version
Git's docs [1] suggestion for adding completions is a bit strange:
you need to get a shell script out of git's source code and source
it locally.

That's what I'd done here, in `git-completions.sh`, but it was already
out of date: ie. it doesn't support completion for the newish
`git switch` and `git restore`

It seems that there are completions available in Ubuntu [2], and when I
tried these they work with newer git commands, so it seems preferable
to use these rather than storing a completions script that I have to
keep up-to-date.

Because this file isn't guaranteed to exist, let's source it
conditionally. Since that leaves the "completion" directory here
misnamed, let's move the `git-prompt.sh` script into "scripts" and
delete ".dotfiles/completion"

[1] https://git-scm.com/book/en/v2/Appendix-A:-Git-in-Other-Environments-Git-in-Bash
[2] https://stackoverflow.com/a/56756511/180500