Install PHP via Ansible
Ideally, PHP would be installed via some sort of version manager like
asdf. However, this is complicated by the fact that the requirements
differe depending on whether you want to use it with Apache or nginx.

In practice, for this reason I rarely tend to install PHP locally but
instead install it in a Vagrant virtual machine on a per-project basis.

So I'm not sure I'll ever need to run this locally, but it feels worth
keeping for reference rather than deleting it, and replacing the shell
script with an Ansible playbook seems worthwhile because a) it's trivial
to convert, b) the Ansible version is more likely to be useful for
provisioning a VM, and c) it gets me one step closer to getting rid of
all the old shell scripts in `setup`