zsh: gitignore .zshrc.zwc
zsh keeps generating this file. It seems to be some sort of compiled
version of .zshrc, though the documentation is pretty sparse. Anyway,
it's a binary file and I *think* it should be ok not to commit it, so
let's ignore it.

At the moment I'm just playing with zsh on a branch while I decide
whether I like it or not, but I'll commit this to main so this file
is always ignored.