Allow "git alias" to accept an argument
`alias` can be run on its own to show all aliases, but with an argument
to show a specific alias.

Let's tweak `git alias` so it works in a similar way: `git alias` lists
all aliases, while, eg. `git alias rb` lists all aliases that have
"rb" in them.

This function is based on this example [1]. However, passing the regex to
`git config --get-regexp` would only search the command aliases, not the
commands themselves. I felt it would be useful to search both the aliases
*and* the aliased commands, so I've tweaked it a bit to grep the output
of `git config --get-regexp ^alias.`

[1] https://stackoverflow.com/a/42125272/180500