Split bash_prompt definition to clarify purpose
I'd like to tidy up how my bash_prompt is set.

The first bit is a bit mysterious, but it sets the window title to the
current working directory.

Because I've switched to double quotes to interpolate SET_WINDOW_TITLE,
I need to change how `__git_ps1` is called: it now needs to be wrapped in
`\$()` - the `$` must be escaped if this is to be evaluated every time.
See [1] for more details.

[1] https://wiki.archlinux.org/index.php/Bash/Prompt_customization#Embedding_commands