Display git mob co-authors on command line
It's easy to forget to `git solo` to reset the co-authors.
Showing co-authors' initials on the command line makes it much easier
to keep track of this.

One downside is that it has a small but noticeable effect on performance:
when I hit "Enter" there's a fraction of a second delay that isn't there
when I'm not checking the initials. It's not bad, but I wonder if there's
any way of avoiding this...