Allow z to use fzf when called without arguments
This was added previously but was no longer working, probably because
.aliases is now sourced before z in .bashrc

Moving this function into .bashrc, alongside where z is initialised
fixes this and should make it more likely that they'll continue to
be run in the right order in future.

I've also moved the `z` stuff further down .bashrc to maintain the
approximately alphabetical order in which everything else
gets initialised.