.gitconfig: add "fixup" and "squash" aliases
We can use the super-handy `fzf` here to allow quick, interactive
selection of fixup / squash targets.

The pattern here is the same in both cases:

1. Pipe `git log --oneline` into fzf
2. When a line is selected, use `awk` to pick out the hash
3. Pass the hash to `git commit` using `xargs`

In the case of `squash`, because it opens Vim we need to call `xargs`
with the `-o` flag, otherwise Vim gets confused [1].

This is based on an idea I spotted in [this .gitconfig file][2]
I'm not sure I want the rebase/autosquash to happen automatically as
part of this command, but using fzf this way to select the commit
seemed nice.

[1] https://stackoverflow.com/a/3852702/180500
[2] https://github.com/gko/dotfiles/blob/296901dea9e757576448fa91ff10a534f9d535da/.gitconfig#L35