Set Git's include.path via Ansible
Previously this was done in `bootstrap.sh`, but there's no particular
reason to do it there. It seems tidier to set it via Ansible, alongside
the other git tasks.

Note that using Ansible's `{{ home }}` variable rather than `~` is needed
to make this task idempotent: when I set the value to
`~/.dotfiles/.gitconfig` the task was always marked "changed". I think
this was because the `~` was being substituted for the full path to the
home directory. Using `{{ home }}` instead seems to avoid this and
preserves idempotence: the first run is marked "changed", and subsequent
runs are not.