Don't specify an ignoreRevsFile for git blame
Adding this seemed like a good idea at the time - I thought I could
specify a conventional filename for an ignoreRevs file that I can then
use in any repo without further configuration.

The problem is, `git blame` is not happy if this file doesn't exist,
so there's no point in trying to set a convention here - it just causes
problems, so let's stop doing this.