Install pv
A few days after throwing away an old setup script in which I used to
install `pv`, I found myself needing it (to view progress when copying
and unzipping a very large zip file). It's only a few kB, so let's just
add it back in.

I'd forgotten to tag this task too, so let's also add a tag.