Remove Docker setup script
It used to be necessary to install Docker Engine separately in WSL.
With WSL2 I believe this isn't necessary any more: if you install Docker
in Windows and opt to use the WSL2-based engine, it will just work in WSL
without the need for installing a bunch of extra stuff.

Even if that's not the case, I'm unlikely to want to run this all again,
so I think it can go.