Don't try to show git mob initials if git mob isn't installed
I only use `git mob` on my work laptop, so I want to add initials to
the prompt in a way that won't cause issues if it's not present.

Testing for the presence of `~/.git-coauthors` (which is created by git
mob) isn't a *perfect* test for the presence of git mob, but it is quick
and simple, and will probably do.