Add .terraformrc
The settings here tell Terraform to cache plugins. I'm still having
trouble running Terraform and this hasn't fully solved it, so I don't
know if I'll keep this in the long term, but I'm committing these changes
for the sake of keeping track of how my machine is currently set up.