Add task to install latest LTS version of Node.js
Previously, Ansible installs `n` with the current LTS version of Node.js
when it's first run, but after that leaves updating Node versions as a
manual task.

However, on reflection it's probably best to update Node via Ansible too,
for the sake of making sure that different machines I work on stay
in sync.

The LTS version of Node.js gets updated once a year [1] so this isn't
something that changes too often, and it probably makes sense to be using
the most recent LTS version by default, and `n` makes it easy to keep
specific projects pinned to a different Node version.

[1] https://nodejs.org/en/about/releases/