Bump gh version to 1.2.1
For some reason this time gh grumbled about not being able to find itself
in /usr/local/bin/gh. Symlinking this to where it actually lives
fixed this.

I *could* now install gh via apt, but because that would involve adding
an apt repository, which has sometimes been a source of annoyance in the
past, and this method still seems to work ok, let's just keep going
with installing it directly for now.