Add "tig" to Highlights section in README
`tig` is great for browsing a git repo's history.
It's also pretty handy for staging and unstaging files.

I'm linking to an article about tig rather than to the tig homepage,
just because I think it does a better job of summarising what tig is
and why you might use it.