Always install asdf
Previously asdf was under the "extra" role. However, it's fairly small
(under 5MB) and seems like a nice way of managing versions (for at least
some languages / tools).

If any other Ansible tasks are going to use asdf to install things
though (as we're already doing for Ruby), we should always install it to
ensure that it's always available.