Add `git reuse` alias for `git commit --reuse-message`
`git commit --reuse-message` (aka `git commit -C`) allows you to create a
commit that copies the message from another commit.

This can be particularly handy when you want to reset an existing commit
and then recommit a tweaked version of it, but retain the old commit's
message. In this case you can commit the updated version with:

    git commit -C HEAD@{1}

...since HEAD@{1} refers to the old commit in the reflog.