Show contents of git stashes by default
The default view of a stash (via, eg. `git stash show stash@{0}`) is a
diffstat (ie. like the output of `git show --stat`).

I generally want to see the actual contents of the stash - this is much
more useful for figuring out whether the stash can be dropped or not.

Fortunately, this can be made the default behaviour by setting
`stash.showPatch` to `true` in .gitconfig [1]

[1] https://www.git-scm.com/docs/git-config#Documentation/git-config.txt-stashshowPatch