.gitconfig: colour git blame output by age
`git blame` has a `--color-by-age` flag for indicating more recently
changed lines in a file. I like this, so let's use it by default.

The default settings are `blue,12 month ago,white,1 month ago,red` [1],
but I prefer a colour scheme that relates age to colour more obviously,
so let's change these so newer lines are bright green, and the oldest are
blue, with duller green for lines of intermediate age.

[1] https://git-scm.com/docs/git-config#Documentation/git-config.txt-colorblamehighlightRecent