Ansible: add private role for non-version-controlled tasks
Sometimes I want to install something (in this case the AWS session
manager plugin [1] that I really just need on one machine and am unlikely
to need elsewhere. It's still useful to create an Ansible task for this
as a record of how I installed it, but to keep my dotfiles tidy let's do
that via a role that is kept out of version control.

[1] https://docs.aws.amazon.com/systems-manager/latest/userguide/session-manager-working-with-install-plugin.html#install-plugin-debian