Add .default-gems file to auto-install bundler and solargraph
The `.default-gems` file, if present, will be used by asdf-ruby to
install a set of gems whenever a new version of Ruby is installed [1]

The Ruby Solargraph plugin in VS Code needs the `solargraph` gem to be
installed. Installing it by default should avoid issues with this plugin
when I change Ruby versions.

Likewise, whenever I install a new version of Ruby I need to install
bundler, so let's add that by default too.

[1] https://github.com/asdf-vm/asdf-ruby#default-gems