Update asdf Ruby plugin before installing Ruby
In order to make sure that the desired version of Ruby is available, it
might be necessary to update the asdf ruby plugin first.

As far as Ansible is concerned this task will always run (ie. Ansible
will always mark it as "changed"). I don't know how to avoid this, but
I don't think it should matter since if the plugin is up-to-date then
running "update" should have no effect - behind the scenes it seems to
just do a `git pull` to update the plugin.

(In other words, `asdf plugin update ruby` is itself idempotent).