Make "Update Ruby plugin" task indicate when it's changed
Because this is just a shell command that effectively ends up doing a
`git pull` each time, it was always being marked "changed". This bothered
me when I introduced it, but at the time I couldn't find any way to
avoid it [1].

However, we can fix this using Ansible's [`changed_when` conditional][2],
which allows you to specify a condition that determines whether or not
the task has caused a change.

Although it's slightly clunky, we can identify a change by the difference
in output when there is a change to be fetched. In this case, if
"From..." is present in stderr, then asdf has updated the plugin. This
should mean the task is only marked "changed" when it actually has
updated the plugin.

[1]: bfdc48470870379177df82222197531bd20e0fab
[2]: https://docs.ansible.com/ansible/latest/user_guide/playbooks_error_handling.html#defining-changed