Add task to install dependencies for Cypress
Getting Cypress to work from Ubuntu in WSL2 requires a few dependencies
to be installed and a few settings tweaked.

The Ansible task here installs the packages required and ensures that
dbus is started with bash (and we update `sudoers` to allow the current
user to `sudo` start dbus without being prompted for a password).

We also need to set the $DISPLAY environment variable to WSL's IP address
in .bashrc

In addition to all this, you'll need to separately install VcXsrv in
Windows. This can be done using winget:

    winget install marha.VcXsrv

...and then launch it via "XLaunch"

I've listed all the packages that Cypress specify as dependencies [1],
but I found I only needed the uncommented ones. This might be because
some of these packages were already installed - I've left them all in
case of issues in future: just uncomment any missing dependencies.

[1]: https://on.cypress.io/required-dependencies