Add Ansible task to install Chrome and Chromedriver
Installing Chrome and Chromedriver in Ubuntu allows us to run
feature specs.

We'll follow more-or-less the process described in [this post][1] and
[this Stack Overflow answer][2].

Note that this depends on an XServer (eg. VcXsrv) installed in Windows.
See the post below for details.

Note also that in future (eg. in Windows 11) this may not be necessary
any more because GUI support will be built into WSL.

Having installed Google Chrome, we can check the version in order to
download the appropriate version of Chromedriver (if required).

Ansible's `get_url` module only downloads the specified file if it
doesn't already exist. As long as we leave the most recently downloaded
version of Chromedriver in /tmp this task will be idempotent.

By registering the result of the "Download Chromedriver" task we can also
ensure that the subsequent "Install Chromedriver" task only happens
when a new version of Chromedriver has been downloaded [2].

[1]: https://www.gregbrisebois.com/posts/chromedriver-in-wsl2/
[2]: https://stackoverflow.com/a/56992668/180500