.gitconfig: set "push.autoSetupRemote" to "true"
`push.autoSetupRemote` is a new config setting in Git [1]. It allows you
to `git push` a local branch that has no upstream branch - in this case
it will automatically set up a remote tracking branch.

Effectively it saves having to do `git push -u origin HEAD` the first
time you push a branch, and having to remember whether or not you've
already done this on subsequent pushes.

[1]: https://twitter.com/JI/status/1546948817462800384