Enable corepack for NodeJS to manage yarn and pnpm
`corepack` is a new experimental feature in NodeJS that allows better
management of alternative package managers (specifically `pnpm` and `yarn`).

It allows you to provide a `packageManager` field in a project's
`package.json` that specifies the package manager (and version) used.

This has a couple of advantages:

1. It prevents you from using `yarn` in a `pnpm` project and vice versa
   (doing so results in a "Usage Error: This project is configured to use
   pnpm" message)
2. It quietly fetches and uses the specified version, even if it differs
   from the globally installed one

This seems useful, especially since I'd like to try using `pnpm` in
future projects.

On a machine with only `npm` installed, I think `corepack enable` is the
only step required here, but note that if you're using a machine that
already has `yarn` or `pnpm` on it you'll need to uninstall them too,
in order to make sure you're using the corepack versions.

I found this straightforward to do with pnpm (`npm uninstall -g pnpm`)
but a little bit more fiddly with yarn, because:

a. It seems I'd installed it via different routes (`npm` + `apt`)
b. Yarn's install folder had also been added to the PATH in `.bashrc`

Uninstalling `yarn` via both `npm` (`npm uninstall -g yarn`) and `apt`
(`sudo apt remove yarn`), and then commenting out the relevant line in
`.bashrc` was required in order to get everything working as expected.