Add "bat" file viewer
"bat" [1]  is a drop-in replacement for "cat", but with added features
like syntax highlighting.

On Ubuntu and Debian, due to a name clash it uses the command "batcat".
However, since I don't use whatever the other "bat" is, let's alias "bat"
to "batcat" so we can just write "bat" instead of "cat".

[1]: https://github.com/sharkdp/bat