Add git-mob to default npm packages
git-mob [1] is a tool that lets you easily add co-authors to your commits.
I use it a lot when pairing, and nowadays it seems it doesn't require the
fiddly extra setup steps that it used to need (ie. setting up a git hook),
and it's a reasonably small package, so let's just install it by default.

[1]: https://github.com/rkotze/git-mob