Add "git mainormaster"
This is a handy little command that tells you whether the current repo
uses "main" or "master". The main use for this is for use in other
commands that need to reference the main branch (eg. checking which
branches have been merged).

Spotted this in [Paul Irish's dotfiles][1], and it was exactly the thing
I'd been looking for recently when trying to write a "git merged" alias.

[1]: https://github.com/paulirish/dotfiles/blob/1d90ebb65cc856e3d97c7cdcce15039e05d65c6e/.gitconfig#L43