Configure Ansible to use Python 3
Previously we were allowing Ansible to decide for itself which version of
Python to use, but this causes warnings in some situations.

Since Python 3 is present on all the OS's I'm currently using these
dotfiles on, let's explicitly tell Ansible to use that.

On Pop!_OS this also required installing the `python3-apt` package (this
seems to be installed by default on Ubuntu).