Add config file specifying theme for bat
The default colour scheme used by `bat` contains a dark grey for line
numbers that isn't very readable with my current terminal colour scheme.

Let's add a config file for `bat` and specify a theme that's more readable.

Note that `.batcatrc` doesn't seem to be the convention (by default
bat's conf file is created at `/etc/bat/config`), but I've named it
this and symlinked it to my home directory for consistency with my
other dotfiles.