Use delta instead of diff-so-fancy
[diff-so-fancy][1] is great, but [delta][2] is newer, more configurable,
and seems to be under more active development.

Let's try it out. I do really like diff-so-fancy's diff style - maybe
just because it's what I'm used to - so for now I'm using delta in
diff-so-fancy emulation mode. `hunk-header-style = omit` tells delta
to leave out the hunk headers (filename + line number), which seem
redundant when diff-so-fancy mode already adds the filename at the top.

Tbh, as this currently stands there is not much  noticeable change: using
delta is more-or-less exactly the same as diff-so-fancy. But there is a
small improvement to word diffs (multiple word diffs per line are now
displayed separately), and I can also now try out delta's other features
(side-by-side diffs, etc).

For future reference: `delta --show-config` is quite useful for seeing
the available config settings and their current values.

One thing to note: delta requires Ubuntu 20.04 or above. I'm currently
still on 18.04 on my home machine, so keeping this in a separate branch
for now.

[1]: https://github.com/so-fancy/diff-so-fancy
[2]: https://github.com/dandavison/delta