Always update refs on rebase
--update-refs is a new rebase feature, introduced in [Git 2.38][1],
which automatically updates any dependent branches following a rebase
to point at the new commits.

For example, if you are working on branch B, based on branch A, and you
rebase with a fixup commit into the shared history of both branches, with
`--update-refs` branch A will also be updated, keeping B and A in the
same relation following the rebase (ie. B will still be based on A, and
the shared history will be updated in both branches), rather than
splitting B from A and keeping A in its previous state.

This seems generally desirable, since it avoids quite a lot of faffing
around to keep multiple branches in sync. So let's enable it by default
for now. If you want to avoid this behaviour you can always rebase with
the `--no-update-refs` flag to opt out.

[1]: https://github.blog/2022-10-03-highlights-from-git-2-38/