Add "resetclock" alias
Sometimes when the host machine has been in sleep mode WSL2's clock
drifts. This is a bit weird and can cause issues when, eg. running
apt-get, as well as incorrect timestamps on Git commits.

The command `sudo hwclock -s` fixes this issue [1]. I don't really need
an alias for this so much as a reminder that this is the command to run.
Adding it as an alias at least improves the chance I'll find it again
in future!

[1]: https://stackoverflow.com/a/66930942/180500