Display some whitespace in diffs in tig
The previous setting ignored all whitespace changes in diffs. However,
this means that a commit that is purely a change of indentation doesn't
have a visible diff. Some purely whitespace changes are meaningful
though (eg. in Sass or HAML).

Switching this to `some` seems to be the equivalent of calling `git diff`
with the `--ignore-space-change` (or `-b`) flag: it ignores changes in
amount of whitespace, but does pick up, eg. whitespace being removed on
a line that is now unindented.

This increases the visibility of certain types of whitespace change
without seeming to add too much noise elsewhere.