Tweak highlight style in vim's visual mode
For some reason in visual mode (Shift+V), Vim was applying a highlight
that was the same colour as the foreground text colour when writing git
commits. This meant that selected text was an unreadable solid block
of colour.

A bit of trial and error (based on the answers to [this Stack Overflow
question][1] led to this setting, which inverts the text colour when
highlighted, making it readable.

[1]: https://stackoverflow.com/questions/3074068/how-to-change-the-color-of-the-selected-code-vim-scheme