Add docker attach alias
This is a command that I can never remember off the top of my head.

It starts docker compose in the background (`-d` flag) and then attaches
the terminal to the `app` container. This is useful for using
interactive debugging tools (eg. `binding.pry` in Rails) in a
dockerised app