Enable git rerere
[rerere][1] (REuse REcorded REsolution) is an optional / experimental
Git feature that allows Git to remember how you resolved a particular
merge conflict and automagically resolve it the same way if you run into
it again.

My mental model of how this works is hazy, but I've been using it for a
while and it seems to work: it occasionally automatically resolves merge
conflicts for me, and so far it's always done this correctly. So let's
enable it.

[1]: https://git-scm.com/docs/git-rerere