tig: always split panels vertically
The default value of `vertical-split` is `auto`, which means on smaller
screens it will split horizontally instead of vertically.

I prefer a vertical split even on small screens. The only tweak needed
is to make it *slightly* wider than the default 50%, in order to
accommodate commit messages comfortably on my laptop.